Admins can approve the verification request
Admins should be able to approve the verification requests. With that implementation, now in the frontend side whenever admin clicks approve for the specific request, this request will be removed from admin's verification requests page and also from the database. isVerified property of the user who sent the request is changed to true.